% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uniprot_to_panel.R
\name{add_uniprot_to_panel}
\alias{add_uniprot_to_panel}
\title{Add UniProt IDs to panel}
\usage{
add_uniprot_to_panel(panel_path)
}
\arguments{
\item{panel_path}{Path to panel file}
}
\value{
None
}
\description{
Add UniProt IDs to panel
}
